Code Heber had used to remove vars from dataset
Added ipynb file the shows how and why I reduced the variables in the data set when I used it for Mini Project 1. This only has the code for "cleaning" that we did in Project 1 plus correlation studies that I used to remove variables in the mini project in order to improve the logistic regression and support vector machine models.